pyproject.toml
[tool.poetry]
name = "video-pipeline"
version = "0.1.0"
description = ""
authors = ["shoot649854 <jojoto9945@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydub = "^0.25.1"
python-dotenv = "^1.0.1"
gradio = "^4.19.2"
soundfile = "^0.12.1"
modelscope = "^1.12.0"
librosa = "^0.10.1"
webrtcvad = "^2.0.10"
speechbrain = "^1.0.0"
speechrecognition = "^3.10.1"
pyttsx3 = "^2.90"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
